# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Portal API",
    "version": "1.0.0",
    "description": "Buy, sell and rent property across Myanmar. Prices are decimal strings in the smallest currency unit (MMK has no subunit), because a kyat sale price can exceed the range where a JSON number stays exact."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create an account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^(\\+?95|0)9\\d{7,10}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 200
                  },
                  "intent": {
                    "default": "SEEKER",
                    "type": "string",
                    "enum": [
                      "SEEKER",
                      "OWNER",
                      "AGENT"
                    ]
                  },
                  "preferredLang": {
                    "default": "my",
                    "type": "string",
                    "enum": [
                      "my",
                      "en"
                    ]
                  }
                },
                "required": [
                  "name",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "preferredLang": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with a password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "identifier",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "preferredLang": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send a one-time code by SMS",
        "description": "Always answers 202 whether or not the number is registered, so it cannot be used to discover which numbers have accounts.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^(\\+?95|0)9\\d{7,10}$"
                  }
                },
                "required": [
                  "phone"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a one-time code for a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^(\\+?95|0)9\\d{7,10}$"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  }
                },
                "required": [
                  "phone",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "preferredLang": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate the refresh token",
        "description": "Replaying an already-rotated token revokes every session in its family, on the assumption that the token leaked.",
        "responses": {
          "200": {
            "description": "Rotated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "phone": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "preferredLang": {
                          "type": "string"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The signed-in user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}